[BI-51] Added bloons script with tiers
Implemented a basic bloons script comprising a Tier struct (controls speed and
color), and a dict of tiers. When a collision is detected with a projectile, we will try to decrement to the next tier; if this is not possible, a 'Pop' method is called (NOT IMPLEMENTED) which should destroy the bloon.